# [SPARK-28023][SQL] Trim the string when cast string type to Boolean/Numeric types