## What changes were proposed in this pull request?

This pr trim the string when cast string type to Boolean/Numeric types for 2 reasons:
1. Our Spark SQL itself is also inconsistent, such as:
```sql
spark-sql> select cast('  1 ' as int), cast(' 1 ' as float);
NULL	1.0
```
2. The mainstream database trims the string when cast string type to Boolean/Numeric types:
**PostgreSQL**:
![image](https://user-images.githubusercontent.com/5399861/59488906-65d10c80-8eb3-11e9-9107-e7d8e4bee3c3.png)
**Teradata**:
![image](https://user-images.githubusercontent.com/5399861/59488767-fd822b00-8eb2-11e9-81f6-f0e6a23f079b.png)
**Oracle:**
![image](https://user-images.githubusercontent.com/5399861/59489315-661dd780-8eb4-11e9-9008-6a483632ad74.png)
**DB2**:
![image](https://user-images.githubusercontent.com/5399861/59500298-beac9f00-8ecb-11e9-9b5b-b042e93d6fc0.png)
**Vertica**:
![image](https://user-images.githubusercontent.com/5399861/59489468-daf11180-8eb4-11e9-8b79-70204ad40560.png)
**SQL Server**:
![image](https://user-images.githubusercontent.com/5399861/59489957-1c35f100-8eb6-11e9-8a67-0859eaeaac58.png)
**MySQL**:
![image](https://user-images.githubusercontent.com/5399861/59489760-a6318a00-8eb5-11e9-981d-906646a29c36.png)
**Hive** fixed this issue by [HIVE-17782](https://issues.apache.org/jira/browse/HIVE-17782):
![image](https://user-images.githubusercontent.com/5399861/59560134-46ff8100-903f-11e9-9817-e2b72995c08a.png)


## How was this patch tested?

unit tests