@MaxGekk 2019-06-15
Just in case, how about to introduce new parameter for `toInt`, `toLong` and etc. (to existing or new methods with additional parameter to control trimming). Trimming inside of loops of those methods shouldn't impact on performance so much, and shouldn't produce additional garbage.