@MaxGekk 2019-06-15
> If this is a correctness issue, we can't put it under a flag, right?

Is it really correctness issue? Is the correct behavior described somewhere in Spark's docs?

> we wouldn't want the behavior to vary with the flag.

The flag would be useful to restore performance in the cases when user's input doesn't contain spaces.

> How would people generally find this, etc.

As any other flags - in docs and in the migration guide.

> The overhead of trimming should be trivial compared to parsing.

The statement should be confirmed by benchmarks, isn't it?