@MaxGekk 2019-06-15
As I can see, `trim()` always does a copy of the original string:
https://github.com/apache/spark/blob/17781d75308c328b11cab3658ca4f358539414f2/common/unsafe/src/main/java/org/apache/spark/unsafe/types/UTF8String.java#L543