@wangyum 2019-06-16
@dongjoon-hyun 
Trim the string when casting string to date/timestamp is fixed by #22943.
Partition columns and non-partition columns are the same code path.
@srowen 
Hive fixed `cast string to numeric types with regards to leading/trailing spaces` by [HIVE-17782](https://issues.apache.org/jira/browse/HIVE-17782):