@kiszk 2019-06-16
I agree with @srowen 's opinion regarding performance. When we add a short-cut to avoid copy, we could minimize performance degradation.