@MaxGekk 2019-06-16
When I talk about a flag, I meant a flag for `toLong`, `toInt` and etc. methods as well, see https://github.com/apache/spark/pull/24872#issuecomment-502394796 . Spark implementation is a copy of Hive's: https://github.com/apache/spark/blob/17781d75308c328b11cab3658ca4f358539414f2/common/unsafe/src/main/java/org/apache/spark/unsafe/types/UTF8String.java#L1026
where input trimming is integrated to the parsing functions:
https://github.com/apache/hive/blob/master/serde/src/java/org/apache/hadoop/hive/serde2/lazy/LazyLong.java#L130-L137

I just want to say that the fix (or a feature if it is not defined by SQL standard) should be made in `UTF8String` instead of `Cast`.

Speaking about SQL config, let's consider 2 use cases:
- User's input doesn't have spaces. In that case, we just introduce additional overhead of trimming strings.
- User's input does have spaces.
  * And the user aware of them and call `trim()` explicitly. In that case, we just add additional implicit trimming which cause additional overhead in current implementation. No benefits too but if it is a critical place in user's code, the user has to spend time in troubleshooting the performance problems.
  * If `trim()` is not called explicitly, the behavior is going to be changed, actually. Here, I think, the SQL flag (in `spark.sql.legacy` namespace) could be useful to restore previous behavior.

Additional thoughts, if we wrap `Cast`'s string input by the expression `StringTrim`, we could eliminate double trimming by folding `StringTrim`.