@MaxGekk 2019-06-16
> We can at least bring the overhead of the common case down a lot, but it's already very small.

I worry mostly regarding the garbage the `trim()` produces. That could impact on GC pauses.

> If it's not clear ... from a standards perspective, then I'd say don't make this change at all.

Let's confirm that this is required by the standard. At the moment, the trimming seems just a nice feature.