@kiszk 2019-06-16
If we would introduce an optimization to return `this` when no space, `trim()` will generate no garbage for the common case.