@srowen 2019-06-16
See https://github.com/apache/spark/pull/24884